+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.1.0] - 2024-10-03
 ### Added
 - `SealOptions`, an interface for directly creating and sealing anonymous files outside the context
   of executing them
@@ -37,7 +39,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Everything!
 
-[Unreleased]: https://github.com/iliana/pentacle/compare/v1.0.0...HEAD
+[Unreleased]: https://github.com/iliana/pentacle/compare/v1.1.0...HEAD
+[1.1.0]: https://github.com/iliana/pentacle/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/iliana/pentacle/compare/v0.2.0...v1.0.0
 [0.2.0]: https://github.com/iliana/pentacle/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/iliana/pentacle/compare/v0.1.0...v0.1.1